[stage1-liveness] welford m_block persist-cap + fp32 stat accumulation
m_block persist-cap (_m_block_cap): caps a full-width reduction's M_BLOCK by the
per-program register budget so a huge-M grid-size autotuner_min raise can't force an
occupancy-starving M_BLOCK; only ever LOWERS an over-raised floor (welford huge-M 16->4/2/1,
rms/layer_norm unchanged). welford fp32 +13% geomean (up to +37.5%), 11 welford cells changed,
728 byte-identical.

welford fp32 stat accumulation: the per-chunk sum_x/sum_x2/chunk*chunk run on fp32 chunk
(the online combine acc_* were already fp32) -> bf16 0.082->0.0156 PASS, fp16 NaN->0.0020 PASS
at full chunk width; HBM load stays half, output casts back to x.dtype.

+++ examples/welford.py
@@ -48,7 +48,12 @@ def welford(
         acc_m2 = torch.zeros_like(acc_cnt)
 
         for tile_n in hl.tile(n):
-            chunk = x[tile_m, tile_n]
+            # Accumulate the per-chunk statistics in fp32 even for half-precision input: the
+            # bf16/fp16 `sum(chunk)` and especially `chunk * chunk` (squaring a half) drift with
+            # chunk width (and overflow to NaN at fp16), so the online combine's fp32 acc_* are
+            # not enough on their own. The HBM load stays half (traffic unchanged); this is an
+            # in-register promote. Output still casts back to x.dtype below.
+            chunk = x[tile_m, tile_n].to(torch.float32)
             # Count of VALID columns (the divisor): use the true valid count, not the constexpr
             # tile width (over-counts last-tile padding). Cast the mask to int32 BEFORE summing —
             # a bool operand gives CuTe a saturating accumulator (-> NaN).
